% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{cmp_proportion}
\alias{cmp_proportion}
\title{plot proportions of a set of categories}
\usage{
cmp_proportion(
  ti,
  xtitle = "",
  ytitle = "",
  xangle = 0,
  margin.l = 0.1,
  legend.pos = "top.center.out",
  legend.dir = "h",
  legend.title = "",
  barwidth = 0.8,
  oneline = F,
  expand.x = c(0.01, 0.01),
  expand.y = c(0.01, 0.04),
  pal = "Pastel2",
  nc = 5
)
}
\description{
plot proportions of a set of categories
}
