% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome.R
\name{flattern_gcoord}
\alias{flattern_gcoord}
\title{flattern 2D genome coordinates tibble[chrom, pos] to 1D coordinates}
\usage{
flattern_gcoord(ti, tz)
}
\description{
flattern 2D genome coordinates tibble[chrom, pos] to 1D coordinates
}
