% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome.R
\name{flattern_gcoord_prepare}
\alias{flattern_gcoord_prepare}
\title{prepare flattern 2D genome coordinates}
\usage{
flattern_gcoord_prepare(t_size, gap = 20000000)
}
\description{
prepare flattern 2D genome coordinates
}
