% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sra.R
\name{get_sra_meta}
\alias{get_sra_meta}
\title{Query SRA using BioProject ID(s) and return meta table of Runs}
\usage{
get_sra_meta(acc, yid)
}
\description{
Query SRA using BioProject ID(s) and return meta table of Runs
}
