% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{go_enrich_genesets}
\alias{go_enrich_genesets}
\title{Perform GO enrichment analysis using hypergeometric test}
\usage{
go_enrich_genesets(tgs, pval.cutoff = 0.05)
}
\description{
Perform GO enrichment analysis using hypergeometric test
}
