% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{go_enrich_gosets}
\alias{go_enrich_gosets}
\title{Perform GO enrichment analysis using hypergeometric test}
\usage{
go_enrich_gosets(gids, tgo. = tgo, pval.cutoff = 0.05,
  srcs = c("uniprot.plants", "arabidopsis", "corncyc", "tfdb",
  "Interproscan5"))
}
\description{
Perform GO enrichment analysis using hypergeometric test
}
