% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{heatmap_hc}
\alias{heatmap_hc}
\title{plot a heatmap with hclust tree shown on left}
\usage{
heatmap_hc(
  ti,
  top = 2,
  bottom = 5,
  text.size = 2.5,
  ratio = 3,
  r.top = 0.5,
  leg = "pairwise similarity",
  tiplab = F
)
}
\description{
plot a heatmap with hclust tree shown on left
}
