% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{hyper_enrich_m}
\alias{hyper_enrich_m}
\title{Perform hypergeometric enrichment test on multiple gene sets}
\usage{
hyper_enrich_m(tg, tgrp, method = "BH", pval.cutoff = 0.05)
}
\description{
Perform hypergeometric enrichment test on multiple gene sets
}
