% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{metrics6}
\alias{metrics6}
\title{Default metrics set for motif prediction}
\usage{
metrics6(
  data,
  truth,
  ...,
  estimate,
  estimator = NULL,
  na_rm = TRUE,
  event_level = yardstick_event_level()
)
}
\description{
Default metrics set for motif prediction
}
