% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{ml1}
\alias{ml1}
\title{train best model using gird search and vfold-CV}
\usage{
ml1(
  ti,
  alg = "rf",
  split_prop = 0.9,
  fold = 10,
  nlevel = 3,
  cpu = 1,
  seed = 26
)
}
\description{
train best model using gird search and vfold-CV
}
