% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{otheme}
\alias{otheme}
\title{flexible custom ggplot theme}
\usage{
otheme(strip.size = 8, margin = c(0.5, 0.5, 0.5, 0.5),
  legend.pos = "", legend.dir = "v", legend.border = F, xticks = F,
  yticks = F, xtitle = F, ytitle = F, xtext = F, ytext = F,
  xgrid = F, ygrid = F, xsize = 8, ysize = 8)
}
\description{
flexible custom ggplot theme
}
