% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{plot_ase}
\alias{plot_ase}
\title{plot ASE allele-frequency boxplot}
\usage{
plot_ase(
  ta,
  th,
  min_rc = 20,
  drc = "h",
  val.col = "black",
  val.fill = "white",
  pal.col = "aaas"
)
}
\description{
plot ASE allele-frequency boxplot
}
