% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{plot_hclust}
\alias{plot_hclust}
\title{plot hclust}
\usage{
plot_hclust(
  tm,
  th,
  min.value = 1,
  pct.exp = 0.5,
  cor.opt = "pearson",
  hc.opt = "ward.D",
  var.lab = "lab",
  var.col = "Genotype",
  lab.size = 2.5,
  expand.x = 0.2,
  pal.col = "aaas"
)
}
\description{
plot hclust
}
