% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{read_go}
\alias{read_go}
\title{Read GO annotation for B73 AGP_v4}
\usage{
read_go(
  src = "uniprot.plants",
  fgo = "~/projects/genome/data/Zmays_B73/61_functional/01.go.tsv"
)
}
\description{
Read GO annotation for B73 AGP_v4
}
