% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive.R
\name{read_gspread}
\alias{read_gspread}
\title{read a generic googlesheet}
\usage{
read_gspread(
  book = "maize_studies",
  sheet = 1,
  cred = "~/.config/google_account_token.json"
)
}
\description{
read a generic googlesheet
}
