% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive.R
\name{read_gspread_master}
\alias{read_gspread_master}
\title{read Google Spreadsheet "maize_studies"}
\usage{
read_gspread_master(
  book = "maize_studies",
  sheet = 1,
  lib = "rnaseq",
  cred = "~/.config/google_account_token.json"
)
}
\description{
read Google Spreadsheet "maize_studies"
}
