% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{rnaseq_cpm}
\alias{rnaseq_cpm}
\title{read one RNA-Seq study result (final)}
\usage{
rnaseq_cpm(yid, diri = "~/projects/rnaseq/data/11_qc", genome = "Zmays_B73")
}
\description{
read one RNA-Seq study result (final)
}
