% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{rnaseq_cpm_raw}
\alias{rnaseq_cpm_raw}
\title{read one RNA-Seq study result (raw)}
\usage{
rnaseq_cpm_raw(
  yid,
  diri = "~/projects/rnaseq/data/11_qc",
  genome = "Zmays_B73"
)
}
\description{
read one RNA-Seq study result (raw)
}
