% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{rnaseq_sample_meta}
\alias{rnaseq_sample_meta}
\title{get RNA-Seq sample meta table}
\usage{
rnaseq_sample_meta(yid = "rn18g", dird = "~/projects/rnaseq/data")
}
\description{
get RNA-Seq sample meta table
}
