% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{sum_stat_tibble}
\alias{sum_stat_tibble}
\title{Get summary statistics for the 2nd+ columns of a tibble}
\usage{
sum_stat_tibble(ti)
}
\description{
Get summary statistics for the 2nd+ columns of a tibble
}
