% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{test_enrich}
\alias{test_enrich}
\title{Generic function to perform hypergeometric/fisher exact test for enrichment}
\usage{
test_enrich(
  hitInSample,
  sampleSize,
  hitInPop,
  popSize,
  deplete = F,
  method = "phyper"
)
}
\description{
Generic function to perform hypergeometric/fisher exact test for enrichment
}
