% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome.R
\name{v3_to_v4}
\alias{v3_to_v4}
\title{Read maize v3 to v4 gene ID mapping}
\usage{
v3_to_v4(
  fm = "~/projects/genome/data2/Zmays_B73/gene_mapping/maize.v3TOv4.geneIDhistory.txt",
  opt = "tibble"
)
}
\description{
Read maize v3 to v4 gene ID mapping
}
