% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{otheme}
\alias{otheme}
\title{flexible custom ggplot theme}
\usage{
otheme(
  margin = c(0.5, 0.5, 0.5, 0.5),
  strip.size = 8,
  strip.style = "dark",
  legend.pos = "",
  legend.dir = "v",
  legend.box = "v",
  legend.vjust = NA,
  legend.title = F,
  legend.border = F,
  panel.border = T,
  panel.spacing = 0.02,
  xticks = F,
  yticks = F,
  xtitle = F,
  ytitle = F,
  xtext = F,
  ytext = F,
  xgrid = F,
  ygrid = F,
  xsize = 8,
  ysize = 8,
  nostrip = T
)
}
\description{
flexible custom ggplot theme
}
